tidy up useGroupByLocation
