+++ page-settings/signal-evaluations.js
@@ -1,3 +1,4 @@
+import { useMemo } from "react";
 import {
   FaRegTimesCircle,
   FaTimes,
@@ -94,3 +95,72 @@ export const getMapIcon = (feature) => {
   );
   return setting.mapIcon;
 };
+
+/**
+ * Group taffic signal study records by location. A location may have many studies,
+ * so we merge them based on their `atd_location_id` key. The map/list displays location
+ * records with info about the related studies.
+ * @param {object} studies - a geojson FeatureCollection of signal studies, fetched from socrata
+ * @returns {object} - a geojson FeatureCollection of `locations`
+ */
+export const useGroupByLocation = (studies) =>
+  useMemo(() => {
+    if (!studies) return;
+    const locationProps = [
+      "atd_location_id",
+      "location_name",
+      "location_status_simple",
+      "council_district",
+    ];
+
+    // create an index of all distinct locations
+    const locationIndex = studies.features.reduce((locationIndex, feature) => {
+      // atd_location_id is the location pk
+      const { atd_location_id } = feature.properties;
+
+      if (atd_location_id && locationIndex[atd_location_id]) {
+        // location already indexed, so append this study to it
+        locationIndex[atd_location_id].properties.studies.push({
+          ...feature.properties,
+        });
+        return locationIndex;
+      }
+
+      // copy the study, which we use as our new location feature
+      const location = {
+        type: "Feature",
+        geometry: feature.geometry,
+        properties: { ...feature.properties },
+      };
+
+      // remove all properties except location attributes
+      Object.keys(feature.properties).forEach((prop) => {
+        if (!locationProps.includes(prop)) {
+          delete location.properties[prop];
+        }
+      });
+
+      // and copy the study feature props - we'll need those in the map popup
+      location.properties.studies = [{ ...feature.properties }];
+      locationIndex[atd_location_id] = location;
+
+      return locationIndex;
+    }, {});
+
+    // unpack the location features
+    const features = Object.values(locationIndex);
+
+    // set location modified date from most-recent study modified date
+    features.forEach((feature) => {
+      const studyDates = feature.properties.studies.map(
+        (study) => study.modified_date
+      );
+      studyDates.reverse();
+      feature.properties.modified_date = studyDates?.[0];
+    });
+    // return locations within a new FeatureCollection
+    return {
+      type: "FeatureCollection",
+      features: [...features],
+    };
+  }, [studies]);
